pyproject.toml
[project]
name = "ai-engineer-mentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.22",
]
